requirements.txt
langchain
langchain-community
openai
pandas
faiss-cpu
streamlit
python-dotenv